switch from positions to vibes (#3306)
This switches us from using positions / formations to indicated what
recipe / formula / blueprint we want to using vibes (combinations of
glyphs) as the key.

Right now this is set up to use "no glyphs" as the "default" -- this can
be done with any number of Cogs as long as they're using the default
glyph. If we want to allow something that's a more complete default
(i.e., it ignores glyphs all together), we'll need to come up with a
different way to look up recipes. I.e., the key space is massive, so
simply populating the dictionary with all possible keys won't work.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211670272176225)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>