Remove visitation counts (#3730)
We added visitation counts as an experiment to see how much they could
help agents learn to explore more. We're not pursuing that path anymore,
and this is a bit of a clunky thing to have in the environment. Removing
as part of complexity reduction.


https://app.asana.com/1/1209016784099267/project/1209041170403474/task/1211937735974209?focus=true

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211941759722069)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>