Reduce the number of short_names (#4150)
This reduces the set of short_names that our policies are using, to
un-clutter `cogames policies`. This still leaves `short_names` as a list
instead of making it the singular `short_name` since this makes
backwards compatibility easier in the future, when we want that.

Manually searched for uses of the removed short_names. There are a bunch
of references to `scripted_unclipping` in docs that need to get cleaned
up for other reasons, but aside from that it looks like we're good.


https://app.asana.com/1/1209016784099267/task/1212009141552109?focus=true

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>
Co-authored-by: Richard Higgins <richard@relh.net>